feat: ScoreComponent
